Moved chunk generation to shaping crate
