+++ crates/shaping/Cargo.toml
@@ -11,6 +11,7 @@ bevy_math = "0.8"
 bevy_utils = "0.8"
 bevy_log = "0.8"
 itertools = "0.10"
+bracket-noise = "0.8"
 
 [dev-dependencies]
 projekto_core = {path = "../core", features = ["auto_load_kinds_descs"]}
