+++ crates/shaping/src/lib.rs
@@ -1,8 +1,10 @@
 use bevy_log::trace;
 use bevy_math::{IVec3, Vec3};
 use bevy_utils::HashSet;
+use bracket_noise::prelude::{FastNoise, FractalType, NoiseType};
 use itertools::Itertools;
 
+use projekto_core::chunk::{ChunkKind, ChunkLight};
 use projekto_core::voxel::{self, ChunkFacesOcclusion, FacesOcclusion};
 
 use projekto_core::{
@@ -82,6 +84,56 @@ pub fn compute_indices(vertex_count: usize) -> Vec<u32> {
     }
 
     res
+}
+
+/**
+ Generates a new chunk filling it with [`ChunkKind`] randomly generated by seeded noise
+*/
+pub fn generate_chunk(local: IVec3) -> Chunk {
+    let mut noise = FastNoise::seeded(15);
+    noise.set_noise_type(NoiseType::SimplexFractal);
+    noise.set_frequency(0.03);
+    noise.set_fractal_type(FractalType::FBM);
+    noise.set_fractal_octaves(3);
+    noise.set_fractal_gain(0.9);
+    noise.set_fractal_lacunarity(0.5);
+    let world = chunk::to_world(local);
+
+    let mut kinds = ChunkKind::default();
+    let mut lights = ChunkLight::default();
+
+    for x in 0..chunk::X_AXIS_SIZE {
+        for z in 0..chunk::Z_AXIS_SIZE {
+            lights.set(
+                (x as i32, chunk::Y_END, z as i32).into(),
+                voxel::Light::natural(voxel::Light::MAX_NATURAL_INTENSITY),
+            );
+
+            let h = noise.get_noise(world.x + x as f32, world.z + z as f32);
+            let world_height = ((h + 1.0) / 2.0) * (chunk::X_AXIS_SIZE * 2) as f32;
+
+            let height_local = world_height - world.y;
+
+            if height_local < f32::EPSILON {
+                continue;
+            }
+
+            let end = usize::min(height_local as usize, chunk::Y_AXIS_SIZE);
+
+            for y in 0..end {
+                // TODO: Check this following biome settings
+                let kind = voxel::Kind::get_kind_with_height_source(end - 1, y);
+
+                kinds.set((x as i32, y as i32, z as i32).into(), kind);
+            }
+        }
+    }
+
+    Chunk {
+        kinds,
+        lights,
+        ..Default::default()
+    }
 }
 
 /**
@@ -285,6 +337,18 @@ fn update_kind_neighborhoods<'a>(world: &mut VoxWorld, locals: impl Iterator<Ite
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn generate_chunk() {
+        let local = (5432, 0, 5555).into();
+        let chunk = super::generate_chunk(local);
+
+        assert!(
+            !chunk.kinds.is_default(),
+            "Generate chunk should should not be default"
+        );
+    }
+
     #[test]
     fn faces_occlusion_occlude_empty_chunk() {
         // Arrange
