+++ src/world/terraformation/genesis.rs
@@ -10,12 +10,11 @@ use bevy::{
     tasks::{AsyncComputeTaskPool, IoTaskPool, Task},
     utils::{HashMap, HashSet},
 };
-use bracket_noise::prelude::{FastNoise, FractalType, NoiseType};
 use futures_lite::future;
 use itertools::Itertools;
 
 use projekto_core::{
-    chunk::{self, Chunk, ChunkKind, ChunkLight},
+    chunk::{self, Chunk},
     voxel::{self},
     VoxWorld,
 };
@@ -55,11 +54,7 @@ fn setup_resources(mut commands: Commands, asset_server: Res<AssetServer>) {
     let vox_world = VoxWorld::default();
     commands.insert_resource(WorldRes(Some(vox_world)));
 
-    let kinds_path = format!(
-        "{}{}",
-        env!("CARGO_MANIFEST_DIR"),
-        "/assets/voxels/kind.ron"
-    );
+    let kinds_path = format!("{}{}", env!("ASSETS_PATH"), "/voxels/kind.ron");
     let descs = voxel::KindsDescs::init(kinds_path);
 
     let atlas = asset_server.load(&descs.atlas_path);
@@ -574,7 +569,7 @@ fn generate_chunks(world: &mut VoxWorld, locals: Vec<IVec3>) -> Vec<IVec3> {
 
     // Before doing anything else, all generated chunks have to be added to world.
     locals.iter().for_each(|&local| {
-        world.add(local, generate_chunk(local));
+        world.add(local, shaping::generate_chunk(local));
     });
 
     // Mark all generated chunks and it's surround as dirty
@@ -590,58 +585,6 @@ fn generate_chunks(world: &mut VoxWorld, locals: Vec<IVec3>) -> Vec<IVec3> {
         .collect();
 
     compute_chunks_internals(world, dirty_chunks)
-}
-
-/**
- Generates a new chunk filling it with [`ChunkKind`] randomly generated by seeded noise
-*/
-fn generate_chunk(local: IVec3) -> Chunk {
-    perf_fn_scope!();
-
-    let mut noise = FastNoise::seeded(15);
-    noise.set_noise_type(NoiseType::SimplexFractal);
-    noise.set_frequency(0.03);
-    noise.set_fractal_type(FractalType::FBM);
-    noise.set_fractal_octaves(3);
-    noise.set_fractal_gain(0.9);
-    noise.set_fractal_lacunarity(0.5);
-    let world = chunk::to_world(local);
-
-    let mut kinds = ChunkKind::default();
-    let mut lights = ChunkLight::default();
-
-    for x in 0..chunk::X_AXIS_SIZE {
-        for z in 0..chunk::Z_AXIS_SIZE {
-            lights.set(
-                (x as i32, chunk::Y_END, z as i32).into(),
-                voxel::Light::natural(voxel::Light::MAX_NATURAL_INTENSITY),
-            );
-
-            let h = noise.get_noise(world.x + x as f32, world.z + z as f32);
-            let world_height = ((h + 1.0) / 2.0) * (chunk::X_AXIS_SIZE * 2) as f32;
-
-            let height_local = world_height - world.y;
-
-            if height_local < f32::EPSILON {
-                continue;
-            }
-
-            let end = usize::min(height_local as usize, chunk::Y_AXIS_SIZE);
-
-            for y in 0..end {
-                // TODO: Check this following biome settings
-                let kind = voxel::Kind::get_kind_with_height_source(end - 1, y);
-
-                kinds.set((x as i32, y as i32, z as i32).into(), kind);
-            }
-        }
-    }
-
-    Chunk {
-        kinds,
-        lights,
-        ..Default::default()
-    }
 }
 
 /**
@@ -964,17 +907,6 @@ mod tests {
         );
         assert!(not_loaded.contains(&local));
         assert!(!world.exists(local), "Chunk should not be added to world");
-    }
-
-    #[test]
-    fn generate_chunk() {
-        let local = (5432, 0, 5555).into();
-        let chunk = super::generate_chunk(local);
-
-        assert!(
-            !chunk.kinds.is_default(),
-            "Generate chunk should should not be default"
-        );
     }
 
     #[test]
